+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added AI assistant (Gemini) feature when writing posts.
 - Added applimode-tool for configuring Applimode projects.
 - Fixed Ref subclasses in Riverpod.
+- Changing screen transition on apple web
 - Improved some docs.
 
 ## Version 0.1.7 (September 12, 2024)